# internal/handlers/service_handler.go
package handlers

import (
	"encoding/json"
	"net/http"
	"quokkaq-go-backend/internal/models"
	"quokkaq-go-backend/internal/services"

	"github.com/go-chi/chi/v5"
)

type ServiceHandler struct {
	service services.ServiceService
}

func NewServiceHandler(service services.ServiceService) *ServiceHandler {
	return &ServiceHandler{service: service}
}

// CreateService godoc
// @Summary      Create a new service
// @Description  Creates a new service for a unit
// @Tags         services
// @Accept       json
// @Produce      json
// @Param        service body models.Service true "Service Data"
// @Success      201  {object}  models.Service
// @Failure      400  {string}  string "Bad Request"
// @Failure      500  {string}  string "Internal Server Error"
// @Router       /services [post]
func (h *ServiceHandler) CreateService(w http.ResponseWriter, r *http.Request) {
	var service models.Service
	if err := json.NewDecoder(r.Body).Decode(&service); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}

	if err := h.service.CreateService(&service); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(service)
}

// GetServicesByUnit godoc
// @Summary      Get services by unit
// @Description  Retrieves all services for a specific unit
// @Tags         services
// @Produce      json
// @Param        unitId path      string  true  "Unit ID"
// @Success      200    {array}   models.Service
// @Failure      500    {string}  string "Internal Server Error"
// @Router       /units/{unitId}/services [get]
func (h *ServiceHandler) GetServicesByUnit(w http.ResponseWriter, r *http.Request) {
	unitID := chi.URLParam(r, "unitId")
	services, err := h.service.GetServicesByUnit(unitID)
	if err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	json.NewEncoder(w).Encode(services)
}

// GetServiceByID godoc
// @Summary      Get a service by ID
// @Description  Retrieves a specific service by its ID
// @Tags         services
// @Produce      json
// @Param        id   path      string  true  "Service ID"
// @Success      200  {object}  models.Service
// @Failure      404  {string}  string "Service not found"
// @Router       /services/{id} [get]
func (h *ServiceHandler) GetServiceByID(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	service, err := h.service.GetServiceByID(id)
	if err != nil {
		http.Error(w, "Service not found", http.StatusNotFound)
		return
	}
	json.NewEncoder(w).Encode(service)
}

// UpdateService godoc
// @Summary      Update a service
// @Description  Updates an existing service
// @Tags         services
// @Accept       json
// @Produce      json
// @Param        id      path      string          true  "Service ID"
// @Param        service body      models.Service  true  "Service Data"
// @Success      200     {object}  models.Service
// @Failure      400     {string}  string "Bad Request"
// @Failure      500     {string}  string "Internal Server Error"
// @Router       /services/{id} [put]
func (h *ServiceHandler) UpdateService(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	var service models.Service
	if err := json.NewDecoder(r.Body).Decode(&service); err != nil {
		http.Error(w, err.Error(), http.StatusBadRequest)
		return
	}
	service.ID = id

	if err := h.service.UpdateService(&service); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}

	json.NewEncoder(w).Encode(service)
}

// DeleteService godoc
// @Summary      Delete a service
// @Description  Deletes a service by its ID
// @Tags         services
// @Param        id   path      string  true  "Service ID"
// @Success      204  {object}  nil
// @Failure      500  {string}  string "Internal Server Error"
// @Router       /services/{id} [delete]
func (h *ServiceHandler) DeleteService(w http.ResponseWriter, r *http.Request) {
	id := chi.URLParam(r, "id")
	if err := h.service.DeleteService(id); err != nil {
		http.Error(w, err.Error(), http.StatusInternalServerError)
		return
	}
	w.WriteHeader(http.StatusNoContent)
}